pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "papermodels"
authors = [{name = "Connor Ferster", email = "connor@structuralpython.com"}]
license = {file = "LICENSE"}
classifiers = ["License :: OSI Approved :: Apache Software License"]
dynamic = ["version", "description"]
dependencies = [
    "shapely >= 2.0.0",
    "matplotlib",
    "numpy >= 2.0.0",
    "pandas >= 2.0.0",
    "pikepdf",
    "sqlalchemy",
    "sqlalchemy-utils",
    "parse",
    "colour",
    "pytest",
    "networkx",
    "numpy",
    "PyNiteFEA >= 0.0.96",
    "pycba",
    "rich",
    "forallpeople"
]


[project.urls]
Home = "https://github.com/structuralpython/papermodels"
